+++ backend/src/main/java/nl/stenden/eindopdracht/model/User.java
@@ -8,7 +8,7 @@ import java.util.Set;
 @Table(name = "user")
 public class User {
 
-    private int id;
+    private Long id;
     private String firstName;
     private String lastName;
     private String email;
@@ -16,14 +16,13 @@ public class User {
     private String password;
     private Set<ProjectGroup> groupList;
 
-
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public int getId(){
+    public Long getId(){
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
